Fix terrain tessellation gaps at chunk boundaries
This change addresses two issues causing visual artifacts at terrain edges:
1. Precision mismatch: uChunkSize was passed as an int to shaders, causing
   truncation errors with non-integer world scales. It is now a float.
2. Inconsistent tessellation: Normal-based boosts in the TCS could differ
   between chunks due to inconsistent normals at the perimeter. Perimeter
   normals are now forced to be perfectly vertical to ensure matching
   subdivision levels.

Modified:
- shaders/terrain.tcs: uChunkSize is now a float.
- shaders/terrain.tes: uChunkSize is now a float.
- src/terrain_render_manager.cpp: Pass uChunkSize as float.
- src/terrain_generator.cpp: Force perimeter normals to vertical.

+++ src/terrain_render_manager.cpp
@@ -515,7 +515,7 @@ namespace Boidsish {
 		shader.setFloat("uTessQualityMultiplier", tess_quality_multiplier);
 		shader.setFloat("uTessLevelMax", 64.0f);
 		shader.setFloat("uTessLevelMin", 1.0f);
-		shader.setInt("uChunkSize", static_cast<int>(chunk_size_ * last_world_scale_));
+		shader.setFloat("uChunkSize", chunk_size_ * last_world_scale_);
 
 		if (clip_plane) {
 			shader.setVec4("clipPlane", *clip_plane);
